refactor: complete ts migration, move code to src and cleanup root app.js (issue #11)
