+++ .gitignore
@@ -2,7 +2,6 @@
 
 node_modules/
 .env
-dist/
 build/
 .DS_Store
